move utilities to utils; fix connect to message
