+++ plugins/atom/lib/utils.js
@@ -23,7 +23,52 @@ const random = (lo, hi) => {
 const getStringPort = () =>
   String(random(60600, 62600));
 
+const createPatches = event =>
+  [
+    {
+      start: {
+        row: event.oldRange.start.row,
+        column: event.oldRange.start.column,
+      },
+      end: {
+        row: event.oldRange.end.row,
+        column: event.oldRange.end.column,
+      },
+      text: '',
+    },
+    {
+      start: {
+        row: event.oldRange.start.row,
+        column: event.oldRange.start.column,
+      },
+      end: {
+        row: 0,
+        column: 0,
+      },
+      text: event.newText,
+    },
+  ];
+
+const eventForBuffer = buffer => (
+  {
+    oldRange: {
+      start: {
+        row: 0,
+        column: 0,
+      },
+      end: {
+        row: 0,
+        column: 0,
+      },
+    },
+    newText: buffer.getText(),
+  }
+);
+
+
 export default {
+  createPatches,
+  eventForBuffer,
   spawnAgent,
   getStringPort,
 };
